apps/skillflex/requirements.txt
Flask
pandas
google-generativeai
python-dotenv
SQLAlchemy
pytest
